==> workbooks/1-1 - Preprocessing USDA.Rmd <==
---
title: "1-1 - Preprocessing USDA"
author: "Julian Barg"
date: "March 21, 2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 1.1 USDA

### Read the full dataset.

We specify the date columns to specify that they are parsed correctly.
```{R read_usda}
library(tidyverse)
library(readxl)

usda_full <- read_excel("../data/USDA_Certified_Wine_2019-01-11.xlsx", na = c(NA, ""), 
                        col_types = c(rep("text", 10), "date", "date", "text", "date", rep("text", 4), "date", rep("text", 4), "date", rep("text", 4),
                                      "date", rep("text", 36), "date", "text"))
head(usda_full)
```

Print row names.
```{R row_names}
colnames(usda_full)
```

### Extract the columns that we actually need.
```{R filter_usda}
usda <- usda_full[c(-1, -2) ,c(4, 5, 1, 6, 13, 14, 15, 28, 29, 30, 33, 35, 38, 41, 43, 46)]
```

```{r usda}
colnames(usda)
```

### Fix address info

Concatenate physical address into one column.
```{R concat_physical_address}
usda$physical_address <- usda %>%
  mutate(street = replace_na(.$`Physical Address: Street 1`, "")) %>%
  mutate(zip = replace_na(.$`Physical Address: ZIP/ Postal Code`, "")) %>%
  mutate(city = replace_na(.$`Physical Address: City`, "")) %>%
  mutate(physical_address = paste0(street, "\n", zip, " ", city)) %>%
  .$physical_address

# Covering the case thate all three columns are empty, and the only thing that ends up in the column are our seperators.
usda$physical_address[usda$physical_address == "\n "] = NA
glimpse(usda$physical_address, 10)
```

Remove old address columns.
```{r find_address_cols}
colnames(usda)
```

```{r remove_address_cols}
usda <- usda[-c(11, 12, 13)]
```

Where physical address is unavailable, substitute mailing address.
```{r concatenate}
usda$mailing_address <- usda %>%
  mutate(street = replace_na(.$`Mailing Address: Street 1`, "")) %>%
  mutate(zip = replace_na(.$`Mailing Address: ZIP/ Postal Code`, "")) %>%
  mutate(city = replace_na(.$`Mailing Address: City`, "")) %>%
  mutate(mailing_address = paste0(street, "\n", zip, " ", city)) %>%
  .$mailing_address

# Covering the case thate all three columns are empty, and the only thing that ends up in the column are our seperators.
usda$mailing_address[usda$mailing_address == "\n "] = NA
glimpse(usda$mailing_address, 10)
```

```{r fill_missing_address}
usda$physical_address[is.na(usda$physical_address)] <- usda$mailing_address[is.na(usda$physical_address)]

sum(is.na(usda$physical_address))
```

Remove unwanted mailing address columns.
```{r find_mailing_cols}
colnames(usda)
```

```{r remove_mailing}
usda <- usda[ -c(11, 12, 13, 15)]

usda <- plyr::rename(usda, c("physical_address" = "address"))
```

### Identify wineries.

We are only interested in certified organic wineries.
```{r identify_wineries}
usda$winery <- FALSE

certified_wine <- str_detect(usda$`Certified Products Under HANDLING Scope`, pattern="wine|Wine")
certified_wine[is.na(certified_wine)] <- FALSE
usda$winery[certified_wine] <- TRUE
sum(usda$winery)
```

### Identify vineyards.

If the observation is also a winery or has has something wine related in the name, and grows grapes, we will assume the grapes are for wine.
```{r identify_vineyards}
usda$vineyard <- FALSE

wine_grapes <- str_detect(usda$`Certified Products Under CROPS Scope`, pattern="wine|Wine")
usda$vineyard[wine_grapes] <- TRUE


vineyard_name <- str_detect(usda$`Operation Name`, pattern="vineyard|Vineyard|wine|Wine")
grapes <-str_detect(usda$`Certified Products Under CROPS Scope`, pattern="grape|Grape")
grapes[is.na(grapes)] <- FALSE

usda$vineyard[((certified_wine | vineyard_name) & grapes)] <- TRUE
```

Drop observations that are not vineyards or wineries.
```{r drop_others}
usda <- usda[usda$winery | usda$vineyard, ]
```

Drop columns that we do not need anymore now
```{r colnames_vineyards_winery}
colnames(usda)
```

```{r drop vineyards_winery_columns}
usda <- usda[ -c(7, 10)]

glimpse(usda)
```

### Handle renamed observations

Clean the column with former names/IDs
```{r clean_former}
former_names_clean <- usda$`Other/Former Names`

former_names_clean <- str_remove(former_names_clean, ".+\\(Former\\sOperation\\sID\\)$")
former_names_clean[former_names_clean==""] <- NA
former_names_clean <- str_remove(former_names_clean, "(^Formerly)")
former_names_clean <- str_remove(former_names_clean, "(^FORMERLY)")
former_names_clean <- str_remove(former_names_clean, "(^Previously)")
former_names_clean <- str_remove(former_names_clean, pattern = ";\\sChanged.+")
former_names_clean <- str_remove(former_names_clean, pattern = "\\sKnow\\sAs\\s")
former_names_clean <- str_remove(former_names_clean, pattern = "(^:?)")
former_names_clean <- str_remove(former_names_clean, pattern = "(^\\s?)")
former_names_clean <- str_remove(former_names_clean, pattern = "(\\s$)?")
# At beginning and end!
former_names_clean <- str_remove(former_names_clean, pattern = "\"")

usda$former_names <- former_names_clean
```

See number of matches

```{r matches}
matches = data.frame()

i=1

for (former_name in usda$former_names) {
  match <- str_which(usda$`Operation Name`, former_name)
  matches <- rbind(matches, c(i, ifelse(length(match) == 0, NA, match)))
  i = i+1
}
names(matches) <- c("index", "match")

sum(!is.na(matches$match))
```

Remove duplicate
Visual inspection of the 7 matches indicates only one real duplicate (matching address, old entry remains in data set). Other cases are probably partial sales of vineyards/wineries. We manually remove the duplicate and clean up the columns.
```{r rempove_duplicate}
usda <- usda[-30, ]
```

```{r list_name_colunmns}
colnames(usda)
```

```{r delete_name_columns}
usda <- usda[ -c(4, 12)]
```

### Handle lapsed certifications
```{r certification_stats}
table(usda$`CROPS Scope Certification Status`)
table(usda$`HANDLING Scope Certification Status`)
```
There are no organizations with lapsed certification in the sample. We can drop the certification column

```{r certification_cols}
colnames(usda)
```

```{r drop_certification}
usda <- usda[ -c(4, 6)]
```

### Rename Columns
```{r colnames_for_rename}
colnames(usda)
```

```{r rename}
usda <- plyr::rename(usda, replace = c("Operation ID" = "id", "Operation Name" = "name", "Certified Name" = "certifier", "Effective Date of CROPS Status" = "vineyard_since", "Effective Date of HANDLING Status" = "winery_since"))
```

### Preview result and save to rds
```{r result}
glimpse(usda)

saveRDS(usda, file = "../preprocessed_data/USDA.rds")
```
