==> workbooks/1-4 - Preprocessing vineyard directory.Rmd <==
---
title: "1-4 Preprocessing vineyard directory"
author: "Julian Barg"
date: "March 23, 2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## 1.3 Lodi sustainable

### Read the full dataset

The overview is more complete, but the details dataframe contains the addresses.
```{r read_lodi}
library(readr)
library(tidyverse)

directory_overview_full <- read_csv("../data/overview_all_vineyards.csv")
directory_details_full <- read_csv("../data/details_all_vineyards.csv")
```

### Rename & drop 
```{r rename}
dir_details <- directory_details_full %>%
  select(-c(Link)) %>%
  plyr::rename(replace = c("Name" = "name", "Date" = "date", "Address" = "address"))

dir_overview <- directory_overview_full %>%
  select(-c(Link)) %>%
  plyr::rename(replace = c("Name" = "name", "Date" = "date"))

glimpse(dir_details)
glimpse(dir_overview)
```

### Retain valid addresses

Some of the addresses only contain rudimentary information. For each organization, we retain the entry with the longest address, to maximize our chances of getting an accurate address. We assume that vineyards & vineries do not relocate.
```{r address}
dir_details <- dir_details %>%
  mutate(len_address = nchar(address)) %>%
  group_by(name) %>%
  arrange(desc(len_address), desc(date)) %>%
  filter(row_number() == 1) %>%
  select(-len_address)

glimpse(dir_details)
```

### Downsample overview to year
```{r downsample_overview}
library(lubridate)

dir_overview <- dir_overview %>%
  mutate(year = year(date)) %>%
  group_by(name, year) %>%
  mutate(last = last(date)) %>%
  filter(date == last) %>%
  ungroup() %>%
  select(-c(date, last))

glimpse(dir_overview)
nrow(dir_overview)
```

### Join overview with addresses from dir_details

```{r match}
directory <- left_join(dir_overview, subset(dir_details, select=c(name, address)))
glimpse(directory)
sum(is.na(directory$address))
length(unique(directory[is.na(directory$address), ]$name))
```

### Save to rds
```{r rds}
saveRDS(directory, "../preprocessed_data/directory.rds")
```
